Move codes from connection.c to postgres_fdw_plus.c and .h
